requirements.txt
langchain>=0.1.0
langgraph>=0.0.60
langchain-openai>=0.1.0
langchain-community>=0.0.20
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings
tiktoken>=0.5.0
langchain-mistralai>=0.1.0